feat: 게시글 CRUD RestDocs 작성

+++ server/src/main/java/com/lookatme/server/board/dto/BoardResponseDto.java
@@ -1,8 +1,8 @@
 package com.lookatme.server.board.dto;
 
 import com.lookatme.server.comment.dto.CommentResponseDtoV2;
-import com.lookatme.server.member.dto.MemberDto;
 import com.lookatme.server.product.dto.BoardProductsResponseDto;
+import com.lookatme.server.member.dto.MemberDto;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -14,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 public class BoardResponseDto {
 
-    private int boardId;
+    private long boardId;
 
     private String userImage;
